Decompose App.tsx with React Query and extracted hooks
Break the 1664-line monolithic App component into domain-specific hooks:

- useAuth: auth status + logout
- useHealth: health polling via React Query (replaces manual setInterval)
- useLayout: sidebar/panel toggle state + localStorage persistence
- useAgents: agent list via React Query cache (SSE updates via setQueryData)
- useSSE: EventSource lifecycle, dispatches to React Query cache
- useMedia: media files via React Query + IntersectionObserver + animations
- useTerminal: WebSocket terminal, xterm setup, reconnection logic

Key improvements:
- SSE agent updates only re-render components subscribed to the agents query,
  not the entire tree (TerminalPane, StatusFooter, etc. are untouched)
- Health polling uses React Query's refetchInterval instead of manual timers
- Media fetching uses React Query with automatic cache invalidation from SSE
- Shared api() utility in lib/api.ts with 401 event dispatching
- App.tsx reduced from 1664 to 595 lines

All 26 E2E tests pass. Type check and production build clean.
